requirements.txt
astroid==2.2.5
autopep8==1.4.4
awscli==1.16.188
beautifulsoup4==4.7.1
botocore==1.12.178
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
docutils==0.14
idna==2.8
isort==4.3.20
jmespath==0.9.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
pyasn1==0.4.5
pycodestyle==2.5.0
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
typed-ast==1.3.5
urllib3==1.25.3
wordcloud==1.5.0
wrapt==1.11.1